refactor: remove duplicate work orders API exports
