+++ src/lib/api.ts
@@ -443,11 +443,3 @@ export {
   normalizeApiError,
 } from './errors';
 export type { ApiError, ApiResponse } from '../../shared/types/http';
-export { workOrdersApi } from './workOrdersApi';
-export type {
-  WorkOrderRecord,
-  WorkOrderStatus,
-  WorkOrderPriority,
-  SaveWorkOrderPayload,
-  PaginatedWorkOrders,
-} from './workOrdersApi';
